+++ datacube/config.py
@@ -6,6 +6,8 @@
 User configuration.
 """
 
+import warnings
+import re
 import os
 from pathlib import Path
 import configparser
@@ -92,6 +94,10 @@ class LocalConfig(object):
         if env:
             if config.has_section(env):
                 self._env = env
+                if not re.match(r'[a-z]+$', self._env):
+                    warnings.warn(f"Configuration environment names like '{self._env}' are deprecated. "
+                    "From datacube 1.9, environment names must start with a lowercase letter and consist of only "
+                    "lowercase letters and digits.", category=DeprecationWarning)
                 # All is good
                 return
             else: